merge master branch

+++ tools/travis_controller.sh
@@ -43,6 +43,33 @@ source "${HERE}/travis/fold.sh"
 source "${HERE}/travis/stage.sh"
 source "${HERE}/travis/shade.sh"
 
+print_system_info() {
+	FOLD_ESCAPE="\x0d\x1b"
+	COLOR_ON="\x5b\x30\x4b\x1b\x5b\x33\x33\x3b\x31\x6d"
+	COLOR_OFF="\x1b\x5b\x30\x6d"
+
+	start_fold "cpu_info" "CPU information"
+	lscpu
+	end_fold "cpu_info"
+
+	start_fold "mem_info" "Memory information"
+	cat /proc/meminfo
+	end_fold "mem_info"
+
+	start_fold "disk_info" "Disk information"
+	df -hH
+	end_fold "disk_info"
+
+	start_fold "cache_info" "Cache information"
+	echo "Maven: $(du -s --si $HOME/.m2)"
+	echo "Flink: $(du -s --si $HOME/flink_cache)"
+	echo "Maven (binaries): $(du -s --si $HOME/maven_cache)"
+	echo "gems: $(du -s --si $HOME/gem_cache)"
+	end_fold "cache_info"
+}
+
+print_system_info
+
 function deleteOldCaches() {
 	while read CACHE_DIR; do
 		local old_number="${CACHE_DIR##*/}"
@@ -56,51 +83,15 @@ function deleteOldCaches() {
 # delete leftover caches from previous builds
 find "$CACHE_DIR" -mindepth 1 -maxdepth 1 | grep -v "$TRAVIS_BUILD_NUMBER" | deleteOldCaches
 
-function getCurrentStage() {
-	STAGE_NUMBER=$(echo "$TRAVIS_JOB_NUMBER" | cut -d'.' -f 2)
-	case $STAGE_NUMBER in
-		(1)
-			echo "$STAGE_COMPILE"
-			;;
-		(2)
-			echo "$STAGE_CORE"
-			;;
-		(3)
-			echo "$STAGE_LIBRARIES"
-			;;
-		(4)
-			echo "$STAGE_CONNECTORS"
-			;;
-		(5)
-			echo "$STAGE_TESTS"
-			;;
-		(6)
-			echo "$STAGE_MISC"
-			;;
-		(7)
-			echo "$STAGE_CLEANUP"
-			;;
-		(*)
-			echo "Invalid stage detected ($STAGE_NUMBER)"
-			return 1
-			;;
-	esac
-
-	return 0
-}
-
-STAGE=$(getCurrentStage)
-if [ $? != 0 ]; then
-	echo "Could not determine current stage."
-	exit 1
-fi
+STAGE=$1
 echo "Current stage: \"$STAGE\""
 
 EXIT_CODE=0
 
 # Run actual compile&test steps
 if [ $STAGE == "$STAGE_COMPILE" ]; then
-	MVN="mvn clean install -nsu -Dflink.forkCount=2 -Dflink.forkCountTestPackage=2 -Dmaven.javadoc.skip=true -B -DskipTests $PROFILE"
+    # We use -Punsafe-mapr-repo since the https version fails on Travis for some reason.
+	MVN="mvn clean install -nsu -Punsafe-mapr-repo -Dflink.convergence.phase=install -Pcheck-convergence -Dflink.forkCount=2 -Dflink.forkCountTestPackage=2 -Dmaven.javadoc.skip=true -B -DskipTests $PROFILE"
 	$MVN
 	EXIT_CODE=$?
 
@@ -109,24 +100,11 @@ if [ $STAGE == "$STAGE_COMPILE" ]; then
         printf "Checking scala suffixes\n"
         printf "==============================================================================\n"
 
-        ./tools/verify_scala_suffixes.sh
+        ./tools/verify_scala_suffixes.sh "${PROFILE}"
         EXIT_CODE=$?
     else
         printf "\n==============================================================================\n"
         printf "Previous build failure detected, skipping scala-suffixes check.\n"
-        printf "==============================================================================\n"
-    fi
-
-    if [ $EXIT_CODE == 0 ]; then
-        printf "\n\n==============================================================================\n"
-        printf "Checking dependency convergence\n"
-        printf "==============================================================================\n"
-
-        ./tools/check_dependency_convergence.sh
-        EXIT_CODE=$?
-    else
-        printf "\n==============================================================================\n"
-        printf "Previous build failure detected, skipping dependency-convergence check.\n"
         printf "==============================================================================\n"
     fi
     
@@ -160,8 +138,19 @@ if [ $STAGE == "$STAGE_COMPILE" ]; then
             # the packing&upload / download&unpacking process
             # by removing files not required for subsequent stages
     
-            # original jars
-            find "$CACHE_FLINK_DIR" -maxdepth 8 -type f -name 'original-*.jar' | xargs rm -rf
+            # jars are re-built in subsequent stages, so no need to cache them (cannot be avoided)
+            find "$CACHE_FLINK_DIR" -maxdepth 8 -type f -name '*.jar' \
+            ! -path "$CACHE_FLINK_DIR/flink-formats/flink-csv/target/flink-csv*.jar" \
+            ! -path "$CACHE_FLINK_DIR/flink-formats/flink-json/target/flink-json*.jar" \
+            ! -path "$CACHE_FLINK_DIR/flink-formats/flink-avro/target/flink-avro*.jar" \
+            ! -path "$CACHE_FLINK_DIR/flink-runtime/target/flink-runtime*tests.jar" \
+            ! -path "$CACHE_FLINK_DIR/flink-streaming-java/target/flink-streaming-java*tests.jar" \
+            ! -path "$CACHE_FLINK_DIR/flink-dist/target/flink-*-bin/flink-*/lib/flink-dist*.jar" \
+            ! -path "$CACHE_FLINK_DIR/flink-dist/target/flink-*-bin/flink-*/opt/flink-table*.jar" \
+            ! -path "$CACHE_FLINK_DIR/flink-dist/target/flink-*-bin/flink-*/opt/flink-python*.jar" \
+            ! -path "$CACHE_FLINK_DIR/flink-connectors/flink-connector-elasticsearch-base/target/flink-*.jar" \
+            ! -path "$CACHE_FLINK_DIR/flink-connectors/flink-connector-kafka-base/target/flink-*.jar" \
+            ! -path "$CACHE_FLINK_DIR/flink-table/flink-table-planner/target/flink-table-planner*tests.jar" | xargs rm -rf
     
             # .git directory
             # not deleting this can cause build stability issues
@@ -197,16 +186,21 @@ elif [ $STAGE != "$STAGE_CLEANUP" ]; then
 	# adjust timestamps to prevent recompilation
 	find . -type f -name '*.java' | xargs touch
 	find . -type f -name '*.scala' | xargs touch
+	# wait a bit for better odds of different timestamps
+	sleep 5
 	find . -type f -name '*.class' | xargs touch
 	find . -type f -name '*.timestamp' | xargs touch
 	travis_time_finish
 	end_fold "adjust_timestamps"
 
-	TEST="$STAGE" "./tools/travis_mvn_watchdog.sh" 300
+	TEST="$STAGE" "./tools/travis_watchdog.sh" 300
 	EXIT_CODE=$?
-else
+elif [ $STAGE == "$STAGE_CLEANUP" ]; then
 	echo "Cleaning up $CACHE_BUILD_DIR"
 	rm -rf "$CACHE_BUILD_DIR"
+else
+    echo "Invalid Stage specified: $STAGE"
+    exit 1
 fi
 
 # Exit code for Travis build success/failure
